module-installs/v1/scripts/requirements.txt
# AI Landscape Monitor — Dependencies
requests>=2.31.0
python-dotenv>=1.0.0
